add db access contrl

+++ config/00_cluster_settings.sh
@@ -8,8 +8,10 @@ current_dir=$(dirname "$BASH_SOURCE")
 DBPATH=$current_dir/../db
 DBNAME=$DBPATH/${CLUSTER_NAME}_${CLUSTER_DOMAIN}_db
 
-ID=`date +%Y%m%d%H%M%S`
-cat <<TEMP_STRUCTURE > /tmp/${ID}_tmpstructure
+if [ ! -e $DBNAME -a ${#1} -eq 0 ]; then
+	sleep 1
+	ID=`date +%Y%m%d%H%M%S`"$RANDOM"
+	cat <<TEMP_STRUCTURE > /tmp/${ID}_tmpstructure
 begin exclusive transaction;
 create table if not exists instances (id INTEGER PRIMARY KEY,instance_id TEXT,prop TEXT,value TEXT);
 create table if not exists cluster (id INTEGER PRIMARY KEY,prop TEXT,value TEXT);
@@ -17,13 +19,13 @@ insert or ignore into cluster (id, prop,value) values (1, 'cluster_name','$CLUST
 insert or ignore into cluster (id, prop,value) values (2, 'cluster_domain','$CLUSTER_DOMAIN');
 insert or ignore into cluster (id, prop,value) values (3, 'instances_num','1');
 end transaction;
+commit;
 TEMP_STRUCTURE
-
-if [ ! -e $DBNAME -a ${#1} -eq 0 ]; then
 	[ ! -e $DBPATH ] && mkdir -p $DBPATH
 	touch $DBNAME
-	sqlite3 $DBNAME < /tmp/${ID}_tmpstructure
+	sqlite3 $DBNAME < /tmp/${ID}_tmpstructure 2>&1>/dev/null || true
+	rm -f /tmp/${ID}_tmpstructure;
 fi
-rm -f /tmp/${ID}_tmpstructure;
+
 set +e
